writeshort

+++ src/main/java/org/traccar/protocol/HuabaoProtocolEncoder.java
@@ -100,7 +100,7 @@ public class HuabaoProtocolEncoder extends BaseProtocolEncoder {
                     data.writeByte(0xa1); // parameter id
                     data.writeByte(0x23); // parameter id
                     data.writeByte(0x00); // parameter value length
-                    data.writeInt(command.getInteger(Command.KEY_DURATION));
+                    data.writeShort(command.getInteger(Command.KEY_DURATION));
                     return HuabaoProtocolDecoder.formatMessage(
                         0x7e, HuabaoProtocolDecoder.MSG_LIGHT, id, false, data, true);
                 case Command.TYPE_BUZZER_ON, Command.TYPE_BUZZER_OFF:
@@ -115,7 +115,7 @@ public class HuabaoProtocolEncoder extends BaseProtocolEncoder {
                     data.writeByte(0xa1); // parameter id
                     data.writeByte(0x24); // parameter id
                     data.writeByte(0x00); // parameter value length
-                    data.writeInt(command.getInteger(Command.KEY_DURATION));
+                    data.writeShort(command.getInteger(Command.KEY_DURATION));
                     return HuabaoProtocolDecoder.formatMessage(
                         0x7e, HuabaoProtocolDecoder.MSG_LIGHT, id, false, data, true);
                 default:
